fixup! [JITLink] Add Hexagon ELF backend
- Move applyFixup() and mask-lookup helpers inline into hexagon.h
  (this function is typically very hot during linking).
- Update PLT relocation comments to reference the GOT/PLT stubs TODO.
